extract Filterg model
