+++ lib/public/views/lhcPeriods/overview/LhcPeriodsModel.js
@@ -15,7 +15,7 @@ import { Observable, RemoteData } from '/js/src/index.js';
 import { getRemoteDataSlice } from '../../../utilities/fetch/getRemoteDataSlice.js';
 import { PaginationModel } from '../../../components/Pagination/PaginationModel.js';
 import { lhcPeriodsFilterDefinition } from './filterDefinition.js';
-
+import { FilterigModel } from './FilteringModel.js';
 /**
  * Model representing handlers for lhc periods page
  *
@@ -37,6 +37,7 @@ export class LhcPeriodsModel extends Observable {
 
         this._lhcPeriods = RemoteData.NotAsked();
         this._filteringModel = new FilterigModel(lhcPeriodsFilterDefinition);
+        this._filteringModel.bubbleTo(this)
     }
 
     /**
@@ -129,59 +130,4 @@ export class LhcPeriodsModel extends Observable {
         this._lhcPeriods = RemoteData.NotAsked();
         this._pagination.reset();
     }
-}
-
-class FilterigModel {
-    constructor(filteringConf) {
-        this._areFiltersVisible = false;
-        this._filteringOptions = Object.fromEntries(
-            Object.entries(filteringConf)
-            .map(([filterTargetName, filterDefinition]) => {
-                const { filter, parser } = filterDefinition;
-                if (!filter && parser || filter && !parser) {
-                    throw new Error(`Invalid filtering configuration for target <${filterTargetName}>`);
-                }
-                return [filterTargetName, { parser, values: null }];
-            })
-        );
-    }
-
-    put(target, values) {
-        this._filteringOptions[target].values = values;
-    }
-
-    get(target) {
-        return this._filteringOptions[target].values;
-    }
-
-    /**
-     * Build filtering search query from using stored filters information
-     * @returns {string} filtering seqrch query
-     */
-    getAsSeqrchQueryParamsString() {
-        return Object.entries(this._filteringOptions)
-            .map(({values, parser}) => parser(values)).join('&');
-    }
-
-    isAnyFilterActive() {
-        return Object.entries(this._filteringOptions).some(({values}) => values?.length > 0 || values);
-    }
-
-    /**
-     * Returns information whether filters component should be visble
-     * @returns {boolean}
-     */
-    get areFiltersVisible() {
-        return this._areFiltersVisible;
-    }
-
-    /**
-     * Toggles whether the filters are shown
-     * @returns {void}
-     */
-    toggleFiltersVisibility() {
-        this._areFiltersVisible = !this._areFiltersVisible;
-        this.notify();
-    }
-
 }
